requirement.txt
Package                  Version
------------------------ -----------
aiohappyeyeballs         2.6.1
aiohttp                  3.11.18
aiosignal                1.3.2
annotated-types          0.7.0
anyio                    4.9.0
appnope                  0.1.3
asttokens                3.0.0
attrs                    25.3.0
certifi                  2025.4.26
charset-normalizer       3.4.2
comm                     0.2.1
dataclasses-json         0.6.7
debugpy                  1.8.11
decorator                5.1.1
defusedxml               0.7.1
distro                   1.9.0
executing                0.8.3
frozenlist               1.6.0
groq                     0.25.0
h11                      0.16.0
httpcore                 1.0.9
httpx                    0.28.1
httpx-sse                0.4.0
idna                     3.10
ipykernel                6.29.5
ipython                  9.1.0
ipython_pygments_lexers  1.1.1
jedi                     0.19.2
jiter                    0.10.0
jsonpatch                1.33
jsonpointer              3.0.0
jupyter_client           8.6.3
jupyter_core             5.7.2
langchain                0.3.25
langchain-community      0.3.24
langchain-core           0.3.61
langchain-groq           0.3.2
langchain-openai         0.3.18
langchain-text-splitters 0.3.8
langsmith                0.3.42
marshmallow              3.26.1
matplotlib-inline        0.1.6
multidict                6.4.4
mypy_extensions          1.1.0
nest-asyncio             1.6.0
numpy                    2.2.6
openai                   1.82.0
orjson                   3.10.18
packaging                24.2
parso                    0.8.4
pexpect                  4.8.0
pip                      25.1
platformdirs             4.3.7
prompt-toolkit           3.0.43
propcache                0.3.1
psutil                   5.9.0
ptyprocess               0.7.0
pure-eval                0.2.2
pydantic                 2.11.4
pydantic_core            2.33.2
pydantic-settings        2.9.1
Pygments                 2.19.1
python-dateutil          2.9.0.post0
python-dotenv            1.1.0
PyYAML                   6.0.2
pyzmq                    26.2.0
regex                    2024.11.6
requests                 2.32.3
requests-toolbelt        1.0.0
setuptools               78.1.1
six                      1.17.0
sniffio                  1.3.1
SQLAlchemy               2.0.41
stack-data               0.2.0
tenacity                 9.1.2
tiktoken                 0.9.0
tornado                  6.4.2
tqdm                     4.67.1
traitlets                5.14.3
typing_extensions        4.13.2
typing-inspect           0.9.0
typing-inspection        0.4.0
urllib3                  2.4.0
wcwidth                  0.2.5
wheel                    0.45.1
yarl                     1.20.0
zstandard                0.23.0
